+++ packages/ui5-proxy-middleware/test/ui5/middleware.test.ts
@@ -152,7 +152,7 @@ describe('middleware', () => {
 
         test('none', async () => {
             await getTestServer(config);
-            expect(ui5ProxySpy).toBeCalledWith(
+            expect(ui5ProxySpy).toHaveBeenCalledWith(
                 expect.objectContaining({}),
                 expect.objectContaining({ secure: true, logLevel: 'info' })
             );
@@ -163,7 +163,7 @@ describe('middleware', () => {
                 ...config,
                 proxy: 'http://proxy.example'
             });
-            expect(ui5ProxySpy).toBeCalledWith(
+            expect(ui5ProxySpy).toHaveBeenCalledWith(
                 expect.objectContaining({ proxy: 'http://proxy.example' }),
                 expect.objectContaining({})
             );
@@ -174,7 +174,7 @@ describe('middleware', () => {
                 ...config,
                 debug: true
             });
-            expect(ui5ProxySpy).toBeCalledWith(
+            expect(ui5ProxySpy).toHaveBeenCalledWith(
                 expect.objectContaining({}),
                 expect.objectContaining({ logLevel: 'debug' })
             );
@@ -185,7 +185,7 @@ describe('middleware', () => {
                 ...config,
                 secure: true
             });
-            expect(ui5ProxySpy).toBeCalledWith(expect.objectContaining({}), expect.objectContaining({ secure: true }));
+            expect(ui5ProxySpy).toHaveBeenCalledWith(expect.objectContaining({}), expect.objectContaining({ secure: true }));
         });
 
         test('secure', async () => {
@@ -193,7 +193,7 @@ describe('middleware', () => {
                 ...config,
                 secure: false
             });
-            expect(ui5ProxySpy).toBeCalledWith(expect.objectContaining({}), expect.objectContaining({ secure: false }));
+            expect(ui5ProxySpy).toHaveBeenCalledWith(expect.objectContaining({}), expect.objectContaining({ secure: false }));
         });
 
         test('directLoad', async () => {
@@ -203,7 +203,7 @@ describe('middleware', () => {
             });
 
             await server.get('/index.html');
-            expect(injectScriptsMock).toBeCalled();
+            expect(injectScriptsMock).toHaveBeenCalled();
         });
 
         test('directLoad:error', async () => {
@@ -239,8 +239,8 @@ describe('middleware', () => {
                 }
             ]);
             await getTestServer(config);
-            expect(loadManifestMock).toBeCalled();
-            expect(ui5ProxySpy).toBeCalledWith(
+            expect(loadManifestMock).toHaveBeenCalled();
+            expect(ui5ProxySpy).toHaveBeenCalledWith(
                 expect.objectContaining({ version: ui5Version }),
                 expect.objectContaining({})
             );
@@ -253,13 +253,13 @@ describe('middleware', () => {
                 }
             ]);
             await getTestServer(config);
-            expect(ui5ProxySpy).toBeCalledWith(expect.objectContaining({ version: '' }), expect.objectContaining({}));
+            expect(ui5ProxySpy).toHaveBeenCalledWith(expect.objectContaining({ version: '' }), expect.objectContaining({}));
         });
 
         test('no manifest.json', async () => {
             rootProjectMock.byGlob.mockResolvedValueOnce([]);
             await getTestServer(config);
-            expect(ui5ProxySpy).toBeCalledWith(expect.objectContaining({ version: '' }), expect.objectContaining({}));
+            expect(ui5ProxySpy).toHaveBeenCalledWith(expect.objectContaining({ version: '' }), expect.objectContaining({}));
         });
     });
 });